Configuring according to repl.it
We can run it on repl.it without a main.py file, but but,,, repl.it is just another host, which also has downtime, Even if uptime bots pings the repl continuously, when repl.it goes down for few minutes(which is normal) all the installed modules are removed. Repl.it doesn't seem to auto install modules without a main.py which causes the bot to go offline! Will make a proper structure with the main.py file and other code soon.